refactor: extract plan lifecycle operations from billing domain
Move 6 plan lifecycle functions and 2 error classes from
features/billing/ into features/plans/ where they belong. This
reduces billing/usage.ts by ~210 lines and establishes a clean
domain boundary — billing no longer owns plan creation, status
transitions, or quota enforcement.

Post-relocation hardening includes: safe type coercion in
countPlansContributingToCap, row-count verification in mark
functions, aligned markPlanGenerationFailure signature, tightened
PlanCreationError details type, and tier-resolution documentation.

Changes:
- Create src/features/plans/errors.ts (PlanLimitReachedError, PlanCreationError)
- Create src/features/plans/lifecycle/plan-operations.ts (checkPlanLimit,
  atomicCheckAndInsertPlan, countPlansContributingToCap,
  markPlanGenerationSuccess, markPlanGenerationFailure, checkPlanDurationCap)
- Re-export moved functions from lifecycle/index.ts barrel
- Migrate all 11 consumers to new import paths
- Relocate duration-caps test to tests/unit/plans/
- Remove extracted code from billing/errors.ts and billing/usage.ts
