+++ src/features/billing/usage.ts
@@ -1,12 +1,9 @@
 import { getDb } from '@/lib/db/runtime';
 import { learningPlans, usageMetrics, users } from '@/lib/db/schema';
 import { logger } from '@/lib/logging/logger';
-import type { PdfContext } from '@/features/pdf/context.types';
 import { and, eq, sql } from 'drizzle-orm';
 
 import {
-  PlanCreationError,
-  PlanLimitReachedError,
   UsageMetricsLoadError,
   UsageMetricsLockError,
   UserNotFoundError,
@@ -94,25 +91,6 @@ export async function resolveUserTier(
 
 // Internal alias for backward compatibility
 const getUserTier = resolveUserTier;
-
-/**
- * Check if user can create more plans
- * @returns true if user can create more plans, false otherwise
- */
-export async function checkPlanLimit(
-  userId: string,
-  dbClient: DbClient = getDb()
-): Promise<boolean> {
-  const tier = await getUserTier(userId, dbClient);
-  const limit = TIER_LIMITS[tier].maxActivePlans;
-
-  if (limit === Infinity) {
-    return true;
-  }
-
-  const currentCount = await countPlansContributingToCap(dbClient, userId);
-  return currentCount < limit;
-}
 
 /**
  * Check if user can use regenerations this month
@@ -298,157 +276,6 @@ export async function getUsageSummary(
       limit: limits.monthlyExports,
     },
   };
-}
-
-/**
- * Count the number of plans that currently consume the user's plan quota.
- *
- * We intentionally combine two disjoint buckets to prevent over- or under-counting:
- *  - Quota-eligible plans: persisted plans that should count toward the cap
- *    (e.g. finalized/ready or manually created). These have `isQuotaEligible = true`.
- *  - In-flight generations: plans that are currently being generated so that
- *    concurrent POSTs cannot exceed the cap while the row is not yet eligible.
- *    We scope this to rows with `generationStatus = 'generating'` AND
- *    `isQuotaEligible = false` to avoid double-counting any plan that might
- *    still have a transient 'generating' status after becoming quota-eligible.
- *
- * This helper accepts a DB handle or a transaction so that callers like
- * `atomicCheckAndInsertPlan` can perform an atomic check-and-insert under a
- * single transaction/lock, ensuring correctness under concurrency.
- */
-async function countPlansContributingToCap(
-  dbOrTx: Pick<DbClient, 'select'>,
-  userId: string
-): Promise<number> {
-  // Use a single query with filtered aggregates to count both buckets
-  const [result] = await dbOrTx
-    .select({
-      count: sql`
-        (
-          count(*) FILTER (WHERE ${learningPlans.isQuotaEligible} = true)
-          +
-          count(*) FILTER (
-            WHERE ${learningPlans.generationStatus} = 'generating'
-              AND ${learningPlans.isQuotaEligible} = false
-          )
-        )::int
-      `,
-    })
-    .from(learningPlans)
-    .where(eq(learningPlans.userId, userId));
-
-  // Return the summed count
-  return (result?.count as number) ?? 0;
-}
-
-/**
- * Atomically check plan limit and insert a new plan to prevent race conditions.
- * This uses a database transaction with row-level locking (SELECT FOR UPDATE)
- * to ensure that concurrent requests cannot exceed the user's plan limit.
- *
- * @param userId - The user's UUID
- * @param planData - The plan data to insert (partial learning plan record)
- * @param dbClient - Database client (defaults to runtime DB with RLS)
- * @returns The inserted plan's ID
- * @throws Error if the user has reached their plan limit
- */
-export async function atomicCheckAndInsertPlan(
-  userId: string,
-  planData: {
-    topic: string;
-    skillLevel: 'beginner' | 'intermediate' | 'advanced';
-    weeklyHours: number;
-    learningStyle: 'reading' | 'video' | 'practice' | 'mixed';
-    visibility: 'private';
-    origin: 'ai' | 'manual' | 'template' | 'pdf';
-    extractedContext?: PdfContext | null;
-    startDate?: string | null;
-    deadlineDate?: string | null;
-  },
-  dbClient: DbClient = getDb()
-): Promise<{ id: string }> {
-  return dbClient.transaction(async (tx) => {
-    // Lock the user row for update to prevent concurrent limit checks
-    // This ensures that only one transaction at a time can check/insert plans for this user
-    const [user] = await tx
-      .select({ subscriptionTier: users.subscriptionTier })
-      .from(users)
-      .where(eq(users.id, userId))
-      .for('update');
-
-    if (!user) {
-      throw new UserNotFoundError(userId);
-    }
-
-    const tier = user.subscriptionTier;
-    const limit = TIER_LIMITS[tier].maxActivePlans;
-
-    // If limit is Infinity (pro tier), skip the check
-    if (limit !== Infinity) {
-      // Count existing plans that count toward quota using shared helper
-      const currentCount = await countPlansContributingToCap(tx, userId);
-
-      if (currentCount >= limit) {
-        throw new PlanLimitReachedError(currentCount, limit);
-      }
-    }
-
-    // Insert the plan within the same transaction (atomic with the checks)
-    const [plan] = await tx
-      .insert(learningPlans)
-      .values({
-        userId,
-        ...planData,
-        generationStatus: 'generating',
-        isQuotaEligible: false,
-      })
-      .returning({ id: learningPlans.id });
-
-    if (!plan) {
-      throw new PlanCreationError();
-    }
-
-    return plan;
-  });
-}
-
-export async function markPlanGenerationSuccess(
-  planId: string,
-  dbClient: DbClient = getDb(),
-  now: () => Date = () => new Date()
-): Promise<void> {
-  const timestamp = now();
-
-  await dbClient
-    .update(learningPlans)
-    .set({
-      generationStatus: 'ready',
-      isQuotaEligible: true,
-      finalizedAt: timestamp,
-      updatedAt: timestamp,
-    })
-    .where(eq(learningPlans.id, planId));
-}
-
-type MarkPlanFailureOptions = {
-  now?: () => Date;
-};
-
-export async function markPlanGenerationFailure(
-  planId: string,
-  dbClient: DbClient = getDb(),
-  options?: MarkPlanFailureOptions
-): Promise<void> {
-  const timestamp = (options?.now ?? (() => new Date()))();
-
-  await dbClient
-    .update(learningPlans)
-    .set({
-      generationStatus: 'failed',
-      isQuotaEligible: false,
-      updatedAt: timestamp,
-    })
-    .where(eq(learningPlans.id, planId));
 }
 
 type AtomicUsageType = 'regeneration' | 'export';
@@ -807,44 +634,6 @@ async function incrementPdfUsageInTx(
       updatedAt: new Date(),
     })
     .where(and(eq(usageMetrics.userId, userId), eq(usageMetrics.month, month)));
-}
-
-const TIER_RECOMMENDATION_THRESHOLD_WEEKS = 8;
-
-type PlanDurationCapResult = {
-  allowed: boolean;
-  reason?: string;
-  upgradeUrl?: string;
-};
-
-export function checkPlanDurationCap(params: {
-  tier: SubscriptionTier;
-  weeklyHours: number;
-  totalWeeks: number;
-}): PlanDurationCapResult {
-  const caps = TIER_LIMITS[params.tier];
-  if (caps.maxWeeks !== null && params.totalWeeks > caps.maxWeeks) {
-    const recommended =
-      params.totalWeeks > TIER_RECOMMENDATION_THRESHOLD_WEEKS
-        ? 'pro'
-        : 'starter';
-    return {
-      allowed: false,
-      reason: `${params.tier} tier limited to ${caps.maxWeeks}-week plans. Upgrade to ${recommended} for longer plans.`,
-      upgradeUrl: '/pricing',
-    };
-  }
-  if (
-    caps.maxHours !== null &&
-    params.weeklyHours * params.totalWeeks > caps.maxHours
-  ) {
-    return {
-      allowed: false,
-      reason: `${String(params.tier)} tier limited to ${String(caps.maxHours)} total hours. Upgrade for more time.`,
-      upgradeUrl: '/pricing',
-    };
-  }
-  return { allowed: true };
 }
 
 export const __test__ = { TIER_LIMITS };
