+++ src/features/plans/lifecycle/adapters/quota-adapter.ts
@@ -6,10 +6,10 @@
 
 import {
   atomicCheckAndIncrementPdfUsage,
-  checkPlanDurationCap,
   decrementPdfPlanUsage,
   resolveUserTier,
 } from '@/features/billing/usage';
+import { checkPlanDurationCap } from '../plan-operations';
 import { normalizePlanDurationForTier } from '@/features/plans/api/shared';
 import type { DbClient } from '@/lib/db/types';
